feat: Move `setParameter` to `ParameterUtil`.
