+++ src/main/java/org/spin/grpc/service/PayrollActionNoticeServiceImplementation.java
@@ -58,6 +58,7 @@ import org.spin.backend.grpc.form.ListPayrollMovementsRequest;
 import org.spin.backend.grpc.form.ListPayrollProcessRequest;
 import org.spin.backend.grpc.form.PayrollActionNoticeGrpc.PayrollActionNoticeImplBase;
 import org.spin.backend.grpc.form.SavePayrollMovementRequest;
+import org.spin.base.db.ParameterUtil;
 import org.spin.base.util.LookupUtil;
 import org.spin.base.util.RecordUtil;
 import org.spin.base.util.SessionManager;
@@ -85,6 +86,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -124,7 +126,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		try {
 			//	SELECT Key, Value, Name FROM ...
 			pstmt = DB.prepareStatement(sql, null);
-			ValueUtil.setParametersFromObjectsList(pstmt, parameters);
+			ParameterUtil.setParametersFromObjectsList(pstmt, parameters);
 
 			//	Get from Query
 			rs = pstmt.executeQuery();
@@ -142,6 +144,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 			}
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			throw new AdempiereException(e);
 		} finally {
 			DB.close(rs, pstmt);
@@ -247,7 +250,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		try {
 			//	SELECT Key, Value, Name FROM ...
 			pstmt = DB.prepareStatement(sql, null);
-			ValueUtil.setParametersFromObjectsList(pstmt, parameters);
+			ParameterUtil.setParametersFromObjectsList(pstmt, parameters);
 
 			//	Get from Query
 			rs = pstmt.executeQuery();
@@ -298,6 +301,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -393,7 +397,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		ResultSet rs = null;
 		try {
 			pstmt = DB.prepareStatement(sql, null);
-			ValueUtil.setParametersFromObjectsList(pstmt, parameters);
+			ParameterUtil.setParametersFromObjectsList(pstmt, parameters);
 			
 			//	Get from Query
 			rs = pstmt.executeQuery();
@@ -410,6 +414,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 			}
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			throw new AdempiereException(e);
 		} finally {
 			DB.close(rs, pstmt);
